requirements.txt
streamlit>=1.40.0
pandas>=2.2.0
openpyxl>=3.1.0
requests>=2.31.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
APScheduler>=3.10.4
reportlab>=4.0.9
python-docx>=1.1.2
python-pptx>=0.6.21
statsmodels>=0.14.0
plotly>=5.18.0
folium>=0.15.0
streamlit-folium>=0.15.0
prophet>=1.1.0
matplotlib>=3.9.0
